upload fix

+++ src/utils/services/artikel.service.js
@@ -7,14 +7,57 @@ const prisma = require('../libs/prisma.config');
 const { getTagByName, createTag } = require('./tag.service');
 
 module.exports = {
-  getAllArtikel: async () => {
-    const artikels = await prisma.artikel.findMany();
+  getAllArtikel: async (filter) => {
+    if (filter) {
+      const artikels = await prisma.$queryRaw`
+      SELECT 
+        a.id, 
+        a.thumbnail, 
+        a.judul, 
+        a.isi, 
+        t.nama as tag, 
+        a.createdAt, 
+        a.updatedAt
+      FROM Artikel a 
+      INNER JOIN Tag t ON a.tag_id = t.id
+      WHERE t.nama = '${filter}' 
+      ORDER BY a.createdAt DESC
+    `;
+
+      return artikels;
+    }
+
+    const artikels = await prisma.$queryRaw`
+    SELECT 
+      a.id, 
+      a.thumbnail, 
+      a.judul, 
+      a.isi, 
+      t.nama as tag, 
+      a.createdAt, 
+      a.updatedAt
+    FROM Artikel a 
+    INNER JOIN Tag t ON a.tag_id = t.id
+    ORDER BY a.createdAt DESC
+  `;
 
     return artikels;
   },
 
   getArtikelbyId: async (id) => {
-    const artikel = await prisma.artikel.findUnique({ where: { id: parseInt(id, 10) } });
+    const artikel = await prisma.$queryRaw`
+      SELECT 
+        a.id, 
+        a.thumbnail, 
+        a.judul, 
+        a.isi, 
+        a.createdAt, 
+        a.updatedAt, 
+        t.nama as tag 
+      FROM Artikel a 
+      INNER JOIN Tag t ON a.tag_id = t.id
+      WHERE a.id = ${id}
+    `;
 
     if (!artikel) {
       return { status: false, message: `Artikel with id ${id} not exist!` };
@@ -47,14 +90,22 @@ module.exports = {
     } catch (error) {
       if (error instanceof Prisma.PrismaClientKnownRequestError) {
         const parse = url.parse(thumbnail);
-        fs.unlink(`uploads/${parse.pathname}`, (err) => {
-          if (err) throw err;
+        fs.exists(`uploads/${parse.pathname}`, (exist) => {
+          if (exist) {
+            fs.unlink(`uploads/${parse.pathname}`, (err) => {
+              if (err) throw err;
+            });
+          }
         });
         return { status: false, code: error.code, meta: error.meta, message: error.message };
       }
       const parse = url.parse(thumbnail);
-      fs.unlink(`uploads/${parse.pathname}`, (err) => {
-        if (err) throw err;
+      fs.exists(`uploads/${parse.pathname}`, (exist) => {
+        if (exist) {
+          fs.unlink(`uploads/${parse.pathname}`, (err) => {
+            if (err) throw err;
+          });
+        }
       });
       throw error;
     }
@@ -83,8 +134,12 @@ module.exports = {
         where: { id: parseInt(id, 10) },
       });
       const parse = url.parse(findArtikel.thumbnail);
-      fs.unlink(`uploads/${parse.pathname}`, (err) => {
-        if (err) throw err;
+      fs.exists(`uploads/${parse.pathname}`, (exist) => {
+        if (exist) {
+          fs.unlink(`uploads/${parse.pathname}`, (err) => {
+            if (err) throw err;
+          });
+        }
       });
 
       return artikel;
@@ -100,9 +155,17 @@ module.exports = {
       },
     });
 
+    if (thumbnail === findArtikel.thumbnail) {
+      return artikel;
+    }
+
     const parse = url.parse(findArtikel.thumbnail);
-    fs.unlink(`uploads/${parse.pathname}`, (err) => {
-      if (err) throw err;
+    fs.exists(`uploads/${parse.pathname}`, (exist) => {
+      if (exist) {
+        fs.unlink(`uploads/${parse.pathname}`, (err) => {
+          if (err) throw err;
+        });
+      }
     });
 
     return artikel;
@@ -117,8 +180,12 @@ module.exports = {
 
     const artikel = await prisma.artikel.delete({ where: { id: parseInt(id, 10) } });
     const parse = url.parse(findArtikel.thumbnail);
-    fs.unlink(`uploads/${parse.pathname}`, (err) => {
-      if (err) throw err;
+    fs.exists(`uploads/${parse.pathname}`, (exist) => {
+      if (exist) {
+        fs.unlink(`uploads/${parse.pathname}`, (err) => {
+          if (err) throw err;
+        });
+      }
     });
 
     return artikel;
